resultsgraph & votebutton compons
